pyproject.toml
[project]
name = "tobb-ocr-rest-api"
version = "1.0.0"
description = "TOBB Ticaret Sicil Gazetesi OCR REST API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    "pdfplumber>=0.10",
    "pytesseract>=0.3.10",
    "Pillow>=10.2",
    "opencv-python-headless>=4.9",
    "structlog>=24.1",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "ruff>=0.3",
    "mypy>=1.8",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
